+++ web_modules/runtime.ts
@@ -8,37 +8,70 @@ interface Window
 
 class Runtime
 {
+    private _initialFetch : boolean;
+
     constructor()
     {
+        this._initialFetch = true;
         this.init();
     }
 
-    private handleStylesheetsFetchEvent:EventListener = this.getStylesheets;
-    private handleScriptFetchEvent:EventListener = this.getScripts;
+    private handleStylesheetsFetchEvent:EventListener = this.getStylesheets.bind(this);
+    private handleScriptFetchEvent:EventListener = this.getScripts.bind(this);
 
     private getStylesheets() : void
     {
-        while (window.stylesheets.length)
-        {
-            let stylesheetFile = window.stylesheets[0];
-            let styleElement:HTMLElement = document.head.querySelector(`style[file="${ stylesheetFile }"]`);
-            if (!styleElement)
+        new Promise((resolve)=>{
+            if (window.stylesheets.length === 0)
             {
-                styleElement = document.createElement('style');
-                styleElement.setAttribute('file', stylesheetFile);
-                document.head.appendChild(styleElement);
-                const stylesheetUrl = `${ window.location.origin }/automation/styles-${ document.documentElement.dataset.cachebust }/${ stylesheetFile }`;
-                this.fetchFile(stylesheetUrl)
-                .then(response => {
-                    styleElement.innerHTML = response;
-                })
-                .catch(error => {
-                    console.error(error);
-                });
+                resolve();
             }
 
-            window.stylesheets.splice(0, 1);
-        }
+            let fetched = 0;
+            const requestedStylesheets = [...window.stylesheets];
+            const numberOfRequiredFiles = requestedStylesheets.length;
+
+            while (window.stylesheets.length)
+            {
+                const stylesheetFile = window.stylesheets[0];
+                let styleElement:HTMLElement = document.head.querySelector(`style[file="${ stylesheetFile }"]`);
+                if (!styleElement)
+                {
+                    styleElement = document.createElement('style');
+                    styleElement.setAttribute('file', stylesheetFile);
+                    document.head.appendChild(styleElement);
+                    const stylesheetUrl = `${ window.location.origin }/automation/styles-${ document.documentElement.dataset.cachebust }/${ stylesheetFile }`;
+                    this.fetchFile(stylesheetUrl)
+                    .then(response => {
+                        styleElement.innerHTML = response;
+                    })
+                    .catch(error => {
+                        console.error(error);
+                    })
+                    .then(()=>{
+                        fetched++;
+                        if (fetched === numberOfRequiredFiles)
+                        {
+                            resolve();
+                        }
+                    });
+                }
+                else
+                {
+                    fetched++;
+                    if (fetched === numberOfRequiredFiles)
+                    {
+                        resolve();
+                    }
+                }
+
+                requestedStylesheets.splice(0, 1);
+                window.stylesheets.splice(0, 1);
+            }
+        })
+        .then(()=>{
+            // Do something after the stylesheets finish loading
+        });
     }
 
     private fetchFile(url:string) : Promise<string>
@@ -70,12 +103,12 @@ class Runtime
             }
             
             let fetched = 0;
-            const requestedPackages = window.packages;
+            const requestedPackages = [...window.packages];
             const numberOfRequiredFiles = requestedPackages.length;
 
             while (requestedPackages.length)
             {
-                let file = requestedPackages[0];
+                const file = requestedPackages[0];
                 let element:HTMLElement = document.head.querySelector(`script[file="${ file }"]`);
                 if (!element)
                 {
@@ -97,6 +130,14 @@ class Runtime
                         reject(error);
                     });
                 }
+                else
+                {
+                    fetched++;
+                    if (fetched === numberOfRequiredFiles)
+                    {
+                        resolve();
+                    }
+                }
 
                 requestedPackages.splice(0, 1);
                 window.packages.splice(0, 1);
@@ -113,12 +154,12 @@ class Runtime
             }
 
             let fetched = 0;
-            const requestedModules = window.modules;
+            const requestedModules = [...window.modules];
             const numberOfRequiredFiles = requestedModules.length;
 
             while (requestedModules.length)
             {
-                let file = requestedModules[0];
+                const file = requestedModules[0];
                 let element:HTMLElement = document.head.querySelector(`script[file="${ file }"]`);
                 if (!element)
                 {
@@ -140,6 +181,14 @@ class Runtime
                         reject(error);
                     });
                 }
+                else
+                {
+                    fetched++;
+                    if (fetched === numberOfRequiredFiles)
+                    {
+                        resolve();
+                    }
+                }
 
                 requestedModules.splice(0, 1);
                 window.modules.splice(0, 1);
@@ -156,12 +205,12 @@ class Runtime
             }
 
             let fetched = 0;
-            const requestedComponents = window.components;
+            const requestedComponents = [...window.components];
             const numberOfRequiredFiles = requestedComponents.length;
 
             while (requestedComponents.length)
             {
-                let file = requestedComponents[0];
+                const file = requestedComponents[0];
                 let element:HTMLElement = document.head.querySelector(`script[file="${ file }"]`);
                 if (!element)
                 {
@@ -183,11 +232,28 @@ class Runtime
                         reject(error);
                     });
                 }
+                else
+                {
+                    fetched++;
+                    if (fetched === numberOfRequiredFiles)
+                    {
+                        resolve();
+                    }
+                }
 
                 requestedComponents.splice(0, 1);
                 window.components.splice(0, 1);
             }
         });
+    }
+
+    private scriptsCallback() : void
+    {
+        if (this._initialFetch)
+        {
+            this._initialFetch = false;
+            console.log('Initial scripts have finished loading');
+        }
     }
 
     private async getScripts()
@@ -197,6 +263,7 @@ class Runtime
             await this.getPackages();
             await this.getModules();
             await this.getComponents();
+            this.scriptsCallback();
         }
         catch (error)
         {